# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AdminLoginRequest {
  password: string;
}

export interface AdminLoginResponse {
  token: string;
}

export interface Cat {
  id: number;
  slug: string;
  name: string;
  kanji: string;
  image: string;
  /** @maxItems 15 */
  images: string[];
  role: string;
  breed: string;
  personality: string[];
  bio: string;
  shortBio: string;
  sortOrder: number;
  active: boolean;
  presentNow: boolean;
}

export interface CatInput {
  slug: string;
  name: string;
  kanji: string;
  image: string;
  /** @maxItems 15 */
  images?: string[];
  role: string;
  breed: string;
  personality: string[];
  bio: string;
  shortBio: string;
  sortOrder?: number;
  active?: boolean;
  presentNow?: boolean;
}

export interface CatPresenceInput {
  presentNow: boolean;
}

export interface Photo {
  id: number;
  url: string;
  caption: string;
  category: string;
  sortOrder: number;
}

export interface PhotoInput {
  url: string;
  caption?: string;
  category?: string;
  sortOrder?: number;
}

export interface SiteConfigEntry {
  id: number;
  key: string;
  value: string;
}

export interface SiteConfigInput {
  value: string;
}

export interface PricingPlan {
  id: number;
  slug: string;
  step?: string;
  title: string;
  titleJa?: string;
  price: string;
  priceNote?: string;
  description?: string;
  includes: string[];
  highlight?: boolean;
  badge?: string;
  cta?: string;
  sortOrder?: number;
}

export interface PricingPlanInput {
  slug: string;
  step?: string;
  title: string;
  titleJa?: string;
  price: string;
  priceNote?: string;
  description?: string;
  includes?: string[];
  highlight?: boolean;
  badge?: string;
  cta?: string;
  sortOrder?: number;
}

export interface Faq {
  id: number;
  category: string;
  question: string;
  answer: string;
  questionJa?: string | null;
  answerJa?: string | null;
  questionZh?: string | null;
  answerZh?: string | null;
  sortOrder: number;
}

export interface FaqInput {
  category: string;
  question: string;
  answer: string;
  questionJa?: string | null;
  answerJa?: string | null;
  questionZh?: string | null;
  answerZh?: string | null;
  sortOrder?: number;
}

export interface Event {
  id: number;
  type: string;
  date: string;
  title: string;
  titleJa?: string;
  summary: string;
  image?: string;
  tag?: string;
  sortOrder: number;
}

export interface EventInput {
  type: string;
  date: string;
  title: string;
  titleJa?: string;
  summary: string;
  image?: string;
  tag?: string;
  sortOrder?: number;
}

export interface GoogleStatus {
  connected: boolean;
  email?: string;
  locationId?: string;
  lastSyncHours?: string;
  lastSyncDescription?: string;
  lastSyncPhotos?: string;
  lastSyncPosts?: string;
  lastSyncWebsiteUrl?: string;
  statusHours?: string;
  statusDescription?: string;
  statusPhotos?: string;
  statusPosts?: string;
  statusWebsiteUrl?: string;
}

export interface GoogleAuthUrl {
  authUrl: string;
}

export interface GoogleLocationInput {
  locationId: string;
  accountId: string;
}

export interface GoogleLocation {
  id: string;
  accountId: string;
  name: string;
  accountName: string;
}

export interface GoogleLocationsResponse {
  locations: GoogleLocation[];
}

export interface GoogleSyncRequest {
  fields: string[];
}

export interface GoogleSyncResult {
  success: boolean;
  error?: string;
  syncedAt?: string;
}

export type GoogleSyncResponseResults = { [key: string]: GoogleSyncResult };

export interface GoogleSyncResponse {
  results: GoogleSyncResponseResults;
}

export interface InstagramStatus {
  connected: boolean;
  username?: string;
  accountId?: string;
  lastSyncAt?: string | null;
  reelCount?: number;
}

export interface InstagramAuthUrl {
  authUrl: string;
}

export interface InstagramReel {
  id: number;
  mediaId: string;
  caption: string;
  thumbnailUrl: string;
  permalink: string;
  postedAt?: string | null;
  cachedAt?: string | null;
}

export interface InstagramSyncResponse {
  syncedCount: number;
  total: number;
}

export interface RequestUploadUrlBody {
  name: string;
  size: number;
  contentType: string;
}

export interface RequestUploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export type GetSiteConfig200 = { [key: string]: string };

export type DisconnectGoogle200 = {
  success: boolean;
};

export type SetGoogleLocation200 = {
  locationId: string;
};

export type DisconnectInstagram200 = {
  success: boolean;
};