# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
        "401":
          description: Unauthorized
  /cats:
    get:
      operationId: listCats
      tags: [cats]
      summary: List all cats
      responses:
        "200":
          description: List of cats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Cat"
    post:
      operationId: createCat
      tags: [cats]
      summary: Create a cat
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CatInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cat"
  /cats/{id}:
    get:
      operationId: getCat
      tags: [cats]
      summary: Get a cat by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cat"
        "404":
          description: Not found
    put:
      operationId: updateCat
      tags: [cats]
      summary: Update a cat
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CatInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cat"
    delete:
      operationId: deleteCat
      tags: [cats]
      summary: Delete a cat
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /cats/{id}/presence:
    patch:
      operationId: updateCatPresence
      tags: [cats]
      summary: Toggle cat presence (in café today)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CatPresenceInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cat"
        "404":
          description: Not found
...
  /faqs:
    get:
      operationId: listFaqs
      tags: [faqs]
      summary: List all FAQs
      responses:
        "200":
          description: List of FAQs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Faq"
    post:
      operationId: createFaq
      tags: [faqs]
      summary: Create a FAQ
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FaqInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Faq"
  /faqs/{id}:
    put:
      operationId: updateFaq
      tags: [faqs]
      summary: Update a FAQ
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FaqInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Faq"
    delete:
      operationId: deleteFaq
      tags: [faqs]
      summary: Delete a FAQ
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List all events/news
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create an event
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}:
    put:
      operationId: updateEvent
      tags: [events]
      summary: Update an event
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    AdminLoginRequest:
      type: object
      properties:
        password:
          type: string
      required:
        - password
    AdminLoginResponse:
      type: object
      properties:
        token:
          type: string
      required:
        - token
    Cat:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        name:
          type: string
        kanji:
          type: string
        image:
          type: string
        images:
          type: array
          items:
            type: string
          maxItems: 15
        role:
          type: string
        breed:
          type: string
        personality:
          type: array
          items:
            type: string
        bio:
          type: string
        shortBio:
          type: string
        sortOrder:
          type: integer
        active:
          type: boolean
        presentNow:
          type: boolean
      required:
        - id
        - slug
        - name
        - kanji
        - image
        - images
        - role
        - breed
        - personality
        - bio
        - shortBio
        - sortOrder
        - active
        - presentNow
    CatInput:
      type: object
      properties:
        slug:
          type: string
        name:
          type: string
        kanji:
          type: string
        image:
          type: string
        images:
          type: array
          items:
            type: string
          maxItems: 15
        role:
          type: string
        breed:
          type: string
        personality:
          type: array
          items:
            type: string
        bio:
          type: string
        shortBio:
          type: string
        sortOrder:
          type: integer
        active:
          type: boolean
        presentNow:
          type: boolean
      required:
        - slug
        - name
        - kanji
        - image
        - role
        - breed
        - personality
        - bio
        - shortBio
    CatPresenceInput:
      type: object
      properties:
        presentNow:
          type: boolean
      required:
        - presentNow
...
    Faq:
      type: object
      properties:
        id:
          type: integer
        category:
          type: string
        question:
          type: string
        answer:
          type: string
        questionJa:
          type: string
          nullable: true
        answerJa:
          type: string
          nullable: true
        questionZh:
          type: string
          nullable: true
        answerZh:
          type: string
          nullable: true
        sortOrder:
          type: integer
      required:
        - id
        - category
        - question
        - answer
        - sortOrder
    FaqInput:
      type: object
      properties:
        category:
          type: string
        question:
          type: string
        answer:
          type: string
        questionJa:
          type: string
          nullable: true
        answerJa:
          type: string
          nullable: true
        questionZh:
          type: string
          nullable: true
        answerZh:
          type: string
          nullable: true
        sortOrder:
          type: integer
      required:
        - category
        - question
        - answer
    Event:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        date:
          type: string
        title:
          type: string
        titleJa:
          type: string
        summary:
          type: string
        image:
          type: string
        tag:
          type: string
        sortOrder:
          type: integer
      required:
        - id
        - type
        - date
        - title
        - summary
        - sortOrder
    EventInput:
      type: object
      properties:
        type:
          type: string
        date:
          type: string
        title:
          type: string
        titleJa:
          type: string
        summary:
          type: string
        image:
          type: string
        tag:
          type: string
        sortOrder:
          type: integer
      required:
        - type
        - date
        - title
        - summary
...